Replace current unhelpful unit tests with a simple import test.
Meaningful tests require the availability of a server and are done
in the context of RSP testing.  The existing tests were both tests
for expected failures when attempting to connect to non-Firefly URLs.
These tests now don't produce an error, as in current versions of
the underlying firefly_client library actual access to external
servers is deferred to the point of an image display.

Note also that tests against outside URLs that Rubin doesn't control
can produce unexpected effects in a production environment and are
undesirable from that perspective.